심준보 : 몬스터 디버프 수정중, HP 게이지 추가, 몬스터 오브젝트 풀 생성, 죽음 상태 전환 및 이펙트 완료, Add : HP gauge, monster object pool, die state transition, Mod : monster debuff

+++ ProjectM/Assets/2. Test/4. SJB/Scripts/Monster_CSV/MonsterData.cs
@@ -8,7 +8,7 @@ public class MonsterData
     public virtual float MonsterMoveSpeed { get; protected set; }
     public virtual int MonsterHP { get; protected set; }
     public virtual int MonsterDamage { get; protected set; }
-    public virtual int MonsterRunSpeed { get; protected set; }
+    public virtual float MonsterRunSpeed { get; protected set; }
 
 
     // 보스 몬스터
